Map DEGTYPE 999 to 'Not Available'
We show trainees as 'incomplete' when they have DEGTYPE for
'999' from HESA. This is a valid entry in HESA for 'Unknown'

We want to map DEGTYPE '999' to 'Not available' for Degree type,
and not show the record as incomplete on our platform.

Also includes the migration file that ensures the data backill.