Update the institution name for given degrees
We noticed that the given degrees have the wrong institution name.
There seems to be only 7 degrees affected by this.
The institution_uuid is correctly assigned, but the name was not
mapped to the expected value.

The expected correct name is: University of St Andrews.

This issue is causing the affected trainee records to show as incomplete.