Store placements in a separate table
This PR introduces a new table called `placements`. We want to
stop relying on the `hesa_students` table for the placements
data.

This PR is also updating the current hesa importer to pull the data
into the new table.

The new logic is setting the foundation for users to manually add
placements in the future.