Fix date range bug
The current bug:

There was a logic error in the code that was setting
the date range to ' ' for all trainees of type EYTS,
in the CSV export, even for trainees that did in fact
have a date range available.

This behaviour was preventing providers to bulk recommend
EYTS trainees.

The fix:

Mark the date range as ' ' only for trainees that have no date range
available.