Don't make DQT API update call when recommending
There is already an API call to the `itt-outcome` endpoint and we should
not need to make a separate call to `update`. It seems that if the calls
happen in the wrong sequence we end up with a dead update job because
you can't update a teacher record that has already got a QTS/EYTS date.

This PR skips the update call.