Attempt to fix problem creating degrees
It's (sometimes) not possible to create a new degree because (it seems)
there is a callback issue between `Degree` and `Trainee`. When a new
`Degree` is saved it touches the `Trainee`. There is also an
`after_touch` callback on Trainee that ends up calling save which
triggers a second `INSERT` on the degrees table.

Do we really need to `save` after touch on a `Trainee`?