Avoid overwriting `Placement#id` during save from `FormStore`.
This fixes a bug that manifests as a double save when adding a new
placement to a registered trainee and hence a unique key error. It
appears that the when we try to combine already saved and stashed state
the stashed state can overwrite the `id` of an existing placement with a
`nil` value.

Also adds a new feature spec since the original one in this area doesn't
trigger the bug for some reason.