Include multiple placement events in timeline
Originally we filtered the timeline to only include the last audit
entry in each set grouped by `request_uuid`. This means that if, say,
you add two placements at a time then you'll only see the last one in
the audit trail. We solve that here by special casing placement audit
entries and returning all entries for a given `request_uuid` if one or
more of those entries relates to a `Placement`.