Make placement CSV header validation more granular
Rather than checking that the CSV headers are exactly as expected, we
ensure that all mandatory headers are present and that there are no
extra headers (outside the expected and optional headers).

Also, update the copy on the bulk update page to reflect that we now
allow up to 5 placements in the bulk upload.