Fix example_data:generate rake task
This task was failing with a duplicate key error on the `Placement#slug`
attribute. This happens due to a side-effect of the validation for
trainees with placements. The validation triggers logic in `PlacementsForm`
that ends up populating the `Trainee#placements` association from the
`FormStore`.

There are a number of un-answered questions here:
- Why does `FormStore` get populated when we are just creating data via
  a rake task (there is no form)?
- Why does the validation for `Trainee` end up mutating the data?

The change in this PR is a workaround to get the rake task working
pending further investigation.