Fix spec regression
To emulate what happens in a JS enabled browser we need to set the
`school_id` field and clear the `school_search` field. In real-life the
`school_search` field doesn't even exist in a JS enabled browser because
the autocomplete code mangles it's name so this seems like a reasonable
thing to do.