Rake file and service to import placements from CSV
We have a CSV file containing HESA IDs and URNs for placements. This commit adds
a rake task and service to import placements on the basis of this file.

The file should be uploaded to the database using the Upload model (via
the admin interface). The rake task can then be run with the ID of the
upload.