Move welcome email trigger to lead school controller
There is logic in the `SendWelcomeEmailService` to skip the email unless
the user is associated with a lead school so it doesn't work immediately
after creating a user. Instead we can trigger the service straight after
adding a lead school to a user. We can call this multiple times because
there's already a check based on the `User#welcome_email_sent_at`
timestamp to prevent multiple emails being sent to the same user.