Move legacy `/api` routes to `/autocomplete`
There were a number of routes that used the `/api` root path and `Api`
module. These service the various autocomplete components that we have
in our frontend. Now that we are building a 'real' API these are
somewhat confusing and out of place so this PR moves them to the
`/autocomplete` root path.

The idea being that this should make it very clear what they are used
for. Note that they've never been used as an external API or documented
as such despite appearances.