User system default page size for user search
Without this the `current_page` of the scope is wrong because
it's applying the system-wide default (which is 30 in dev) and that is
then classing with the `limit` set in `UserSearch` (15).

The system wide default (`Settings.pagination.records_per_page`) varies
between different environments so it makes sense to just use that for
all the things. Currently the system wide limit is used on all the other
sys admin tabs, e.g. schools.