Add 3 new routes for ITT reform
This PR adds these routes:

- Postgraduate fee funded
- Postgraduate salaried
- Undergraduate fee funded

There routes will be 'switched on' at a time (to be determined). I've
added a placeholder service that will work out the logic to determine
which trainees see which routes and when. For now it just returns `true`.

Also included here are some notes and questions about this change.

Note that this SPIKE should not be merged in anything like it's current
form.