Fix regression when adding second placement
The cause of this bug was some recently added validations on the
`Placement` model.

These validations check uniqueness of school name, urn etc. This works
well if the values are `nil` or unique but not with empty strings.

The solution here is to use `allow_blank` rather than `allow_nil` on the
uniqueness validations so that, for example, you have have multiple
placements that have a blank school `name`. This is perfectly normal
because you can fully specify the school via the `school` association in
which case the `name` is not needed.