Refactor Api::DegreeAttributes#duplicates?
When the country is not submitted in the params the locale_code would
default to "uk" by the return value of Api::MapHesaAttributes::Degrees::V01#locale_code.

In combination with the fact that the exists?
query would check with both { locale_code: "uk", country: nil } could result
in inconsistencies in the return value of duplicates?