Implement validation for nationality HESA codes
Invalid codes should trigger a 422 status and return an error message.

The mapping from HESA code to Register nationality values is done in the
`Api::MapHesaAttributes::V01` service class. But there is no model to
validate at this point in the pipeline so I've added validation to the
`Api::NationalityAttributes::V01` class. This makes things somewhat
repetitive but (I think) it keeps the validation logic where we might
expect it to be.