Catch exception when parsing a CSV
Currently, the ValidateCsvRow class is validating the row on initialization. Any errors are added to a messages array, and the valid? method just checks if there are any messages in the array. This makes sure that we do the same thing, even if parsing the row raises an exception. An “unexpected” response message is added to the array and that will be displayed to the user.