Authorize TRN submissions with TraineePolicy#update?
There seems to be an edge case where the provider can create a trainee, but then become unaccredited. In this situation, the provider will fail authorisation when trying to submit for a TRN, because we’re authorising with TraineePolicy#create. I’ve updated this to use the update? method in the policy, because that better describes what we’re trying to do at this point.