Restrict Un-Accredited Providers from Creating New Trainees (#4471)
* Update trainee_policy to check accreditation

This adds a new method to UserWithOrganisationContext called `accredited_provider`, which will check if the user’s provider has `accredited` set to true. We then check this in the new? and create? methods in the trainee policy, since only accredited providers should now be able to do this.

* Update trainee_policy_spec

* Rubocop

* Add feature spec for an unaccredited provider

This adds a scenario to the viewing_the_timeline_spec, where we view the page as an accredited and unaccredited provider. We test that the “Create a trainee record” button is hidden for unaccredited providers

* Remove call to skip_authorization in trainees controller

I’m not sure why this line was added, but I think it’s safer to take this out.

* Remove use of send where it’s not needed

* Authorize TRN submissions with TraineePolicy#update?

There seems to be an edge case where the provider can create a trainee, but then become unaccredited. In this situation, the provider will fail authorisation when trying to submit for a TRN, because we’re authorising with TraineePolicy#create. I’ve updated this to use the update? method in the policy, because that better describes what we’re trying to do at this point.