Switch productiondata to import 2024 Apply apps
We need to set the recruitment cycle to 2024 to get a supply of Apply
applications to productiondata in order to test the integration for the
new academic cycle.

I've also reset the last successful import timestamp for 2024 by
deleting `ApplyApplicationSyncRequest` records. So once we deploy this
hange we _should_ start seeing new Apply applications appear in
productiondata.

Note that the 2 jobs that perform the import run at 1am and 4am.