Migrate to azure-blob Gem From azure-storage-blob (#4494)
* Use azure-blob instead of azure-storage-blob

* Update MalwareScan

* Update malware_scan and fix spec

Fetching a blob with a key is different now than it was using azure-storage-blob. This updates the malware scanning code to use this new API and updates the spec for it.

* Add container name to MalwareScan

The previous version of the azure storage gem didn’t require the name of the container, but the new one does.

* Update MalwareScan error handling

This removes an unnecessary check for the response being a Net::HTTPSuccess. If we get a response at all, without raising an error, then we know we’ve been successful. I’ve also made sure that we catch all possible networking errors that get_blob might raise.