Migrating Lead School to Lead Partner - Part III  (#4551)
* Update lead_school key in en.yml

* Update _scholarships.html.erb

* Rename lead_school in _bursaries partial

* Rename “format_lead_school_string”

* Update lead_schools in _scholarships partial

* Remove unused partial

* Update funding uploads html attribute

Changes lead_school_trainee_summary to lead_partner_trainee_summary.

* Update lead_school_payment_schedule

Updates the html attribute to lead_partner_payment_schedule

* Update FundingUploadsController ivars

Updates @lead_school_trainee_summary to @lead_partner_trainee_summary and @lead_school_payment_schedule to @lead_partner_payment_schedule.

* Update lead_school_trainee_summary funding type

Changes the enum value in the FundingUpload model from lead_school_trainee_summary to lead_partner_trainee_summary.

* Fix broken specs

* Change lead_school_payment_schedule to lead_partner_payment_schedule

* Rename LeadSchoolTraineeSummariesImporter

Renamed to LeadPartnerTraineeSummariesImporter in funding_data_importer_spec.rb

* Fix lead_school_payment_schedules_spec

* Rename LeadSchoolPaymentSchedules

Rename to LeadPartnerPaymentSchedules.
Renames apply to importers and parsers.

* Rename LeadSchoolTraineeSummaries

Renamed to LeadPartnerTraineeSummaries.
Change applies to parsers and importers.

* Rename specs for funding importers and parsers

* Rename FundingUploadsController ivars

Renames @lead_school_payment_schedule to @lead_partner_payment_schedule

* Remove lead_school_user helper methods

These aren’t relevant anymore, we only need the lead_partner_user? helper method where applicable.

* Remove lead_schools_only from school search

Removes use of the lead_only scope on Schools when searching. The option is no longer needed.

* Rubocop

* Fix broken view_spec

This had tests for when the user was a lead school user, which is no longer relevant. We already have tests for when the user is a lead partner user.

* Fix schools_controller_spec

* Remove lead_school names from users controller response

* Remove lead school references from PaymentSchedulesController

* Rubocop

* Remove unused routes from trainee summaries and funding controllers

After checking in the console, the route lead_school_funding_path, for example, no longer exists. So I’ve removed any method in these controllers that try to redirect to it.

* Update payment_schedules_controller

Remove unused routes

* Remove unused @lead_schools ivar from organisations controller

* Remove unused include from PersonasController

* Remove unused LeadSchoolForm and spec

* Update funding_csv_export_path

This updates the method in funding_helper.rb to only worry about the provider path and not the lead_school path, which is no longer used.

* Remove unused lead_school_row helper method

* Rubocop

* Update show_provider method in ApplicationRecordCard View

Removes reference to the lead_school? method on UserWithOrganisationContext which is no longer needed.

* Fix view spec

* Remove reference to lead_school from trainee_about/view.html.erb

* Update trainee_policy and spec

Removes the check for lead_school? on the user and removes the unused lead_school_scope method. Also removes unnecessary tests.

* Remove reference to lead_school in trainee_policy

We were using the reference to see if hide_progress_tag is permitted for lead_school users. But this is no longer relevant.

* Update user policy and spec

Removes reference to lead_school when permitting drafts? or bulk_recommend? But these are no longer relevant.

* Remove unused references to lead_school in trainee policy spec

The only references that I kept in were references to the lead_school record type on lead partner. This will be renamed from lead_school to just school at a later date.

* Remove references to lead_school in user policy spec

* Fix broken feature test

Removes an unnecessary test from viewing_an_existing_trainee_spec. Also removes an unnecessary check for lead_school from home_view.rb which was causing other feature specs to fail.

* Rubocop

* Remove lead_school reference from view_preview.rb

* Remove unused spec from home_view_spec

* Remove lead_school references from view_spec

* Rubocop

* Remove lead_school reference from academic_year_filter_options_spec

* Remove user_has_only_lead_school? from UserWithOrganisationContext

It was only used in the single_organisation method, and is no longer necessary. Also updates the spec.

* Remove lead_school method from UserWithOrganisationContext

After removing all uses of this method across the code, this removes the method itself and the specs for it.

* Fix viewing_an_existing_trainee_spec

* Fix viewing_an_existing_trainee_spec.rb

* Fix trainees controller spec

* Update UserWithOrganisationContext#multiple_organisations

This removes a reference to “user.lead_schools”, which is no longer needed, and updates the spec.

* Update UserWithOrganisationContext#no_organisation?

* Update UserWithOrganisationContext#lead_parter_or_provider

Removes the option to look up a lead school as the organisation.

* Remove lead_school references from user_with_organisation_context_spec

* Remove user_in_lead_school_context from TraineePolicy

* Add lead_partner column to RecommendationsUploadRows

This is added alongside the lead_school column for now, so that both can be kept in sync for a time, until lead_school can be dropped.

* Add data migration for RecommendationsUploadRow

This copies all of the lead_school values into the new lead_partner column.

* Add a before_save callback to RecommendationsUploadRow

This makes sure that lead_school and lead_partner and kept in sync, regardless of which of them might have been changed.

* Correct schema version

* Update callback to use dynamic column names

Now the before_save callback can be used for any column names across different tables.

* Move lead school sync logic to a concern

This means that we can add this logic for any type in the app that needs migrating from lead_school to lead_partner regardless of any variation in the column name.

* Move spec into a concerns folder

* Rubocop

* Add lead_partner_urn column to funding_trainee_summary_rows

* Adds the new lead_partner_urn column.
* Adds a data migration, copying from lead_school_urn to lead_partner_urn.
* Adds the LeadSchoolMigration concern to Funding::TraineeSummaryRows
* Declares that lead_school_urn and lead_partner_urn should be kept in sync.

* Rubocop

* Add lead_partner_urn to Hesa::Student table

* Add the lead_partner_urn column to hesa_students.
* Add a data migration, copying from lead_school_urn into lead_partner_urn.
* Include the LeadSchoolMigratable concern in Hesa::Student
* Declare that lead_school_urn and lead_partner_urn should be kept in sync.

* Refactor use of LeadSchoolMigratable

* Fix spelling error

* Remove unused LeadSchoolUserImport service and specs

* Remove lead_school_users association from user

* Remove CopyLeadSchoolToLeadPartner Service and Rake Task

* Remove lead_school_users and users association from school

* Drop lead_school_users table

* Remove lead_school_users from analytics.yml

* Put back school partial

This reverts commit eea63b4cac5ae6bea6fdb7394a3779b92e488e28.

* Update link in _schools partial

* Make sure school still has lead partner association