Lead School to Lead Partner Part V  (#4598)
* Update school form and school detail view

* Fix edit_schools_spec

* Do not set lead_school attribute in school form

Going forward, we should be asking if a school is a lead partner instead.

* Rename lead_school_urn in ITTRecord

Changes the parser to refer to lead_school_urn as lead_partner_urn. This also updates the CreateFromHesa service and any related specs.

* Revert "Do not set lead_school attribute in school form"

This reverts commit 11159cee55bda29ffc86b11bb96e0144f6f34894.

* Remove lead_school column from School

* Fix service specs

* Fix broken form specs

* Fix view_preview spec

* Fix integration specs

* Fix placement_form_spec

* Fix example data

* Fix broken specs

* Remove lead_school from trainees

Instead, the trainee can now get to a school through the lead_partner. The school model has also been updated to have_many trainees through the lead_partner model.

* Keep the name lead_school for trainees for now

Just using school is ambiguous, since we also have an “employing_school” on trainee. This keeps the name lead_school for now, we can rename it later if needed.

* Fix trainee spec

* Fix find_empty_trainees service

* Fix create_from_csv_row_spec

* Rubocop

* Rename lead_school to lead_partner_school for trainees

* Fix trainee report spec

* Fix view previews

* Fix trainee report spec

* Fix broken specs

* Fix service specs