Further checks for lead partner URN API inputs
If `lead_partner_urn: nil` is input to the API services then we should
treat that as a `nil` value rather than trying to lookup a lead partner
with `urn: nil` as that's clearly not the intention and can cause
serious issues with trainees being associated to random SCITT lead
partners.

Same applies to `lead_partner_ukprn`.