[8024] Add mandatory/optional/conditional info to CSV docs (#5046)
* WIP

* Added `validation` attribute to CSV docs

This is set to either:
- Mandatory
- Optional
- Conditional mandatory

For mandatory fields there should be a link to the code that implements
the rule. This should also be given for Conditional mandatory fields
along with an explanation of the rule.

* Expose `Validation` key in CSV docs

This is really just for testing right now.

* Minor changes to the validation content

Changes to the content suggested by Jen.

* Update app/views/bulk_update/add_trainees/reference_docs/fields.yaml

Co-authored-by: Tom Trentham <1910056+tomtrentham@users.noreply.github.com>

* Correction to format and example for `application_id`

This is just an integer

---------

Co-authored-by: Tom Trentham <1910056+tomtrentham@users.noreply.github.com>