[8564] Specify order for trainee uploads on status page (#5333)
* Specify order for trainee uploads on status page

Order was undefined. Should be in descending order of upload.

* Workaround trainee uploads only shown for current year

Because we only list trainee uploads for the current academic year it's
not helpful to create historic records when testing with
`USE_NEXT_ACADEMIC_YEAR`. We can workaround this by making sure these
tests run at least one month into the academic year.