[8677] GOV.UK rebrand (#5431)
* Upgrade govuk-components to 5.11.1

* Update image assets path for rebrand

* Use correct logo in header

Also remove the header navigation and ensure that header logo links to service home page

* Switch to the rebranded service navigation style

We don't need the MoJ navigation bar any longer.

* Remove unused styles for MoJ navigation bar

* Add app--register-interim-rebranded class for overrides

As part of review feedback it was suggested that we wrap all CSS
adjustments needed to accomodate a navigation bar in the GOV.UK header
in a custom class so that they can easily be removed in future.

* Move service name to service navigation component

* Add links to OrganisationSwitcher component

Styling needs work...

* Reinstate right align style in navigation bar

* Extra links for navigation bar

* Align Support and Sign out links to the right

This doesn't seem to be support OOTB by the service navigation component

* Shorten the support link title

Probably quite important because the navigation links will start
wrapping if they are too long.

* Drop the links in the header

* Modify `Header::View` component to reflect new brand

This no longer has any `items` or a `title` because they no longer go in
the header. It's just the GOV.UK logo.

* Modify `NavigationBar::View` component to reflect new brand

We've dropped the MoJ component styles in favour of the standard GOV.UK
ones now that they are directly supported by the design system.

We also have a service title to test.

* Update spec for OrganisationSwitcher::View

This component now has an extra link for _Organisation settings_.

* Fix the component preview specs

* Fix more feature specs

* Fix feature specs - correction to page object

Was using old MoJ styles

* Correction to `assets_path`

* CSS tweak for organisation switcher on tablets

Avoid misalignment of links and stop floating them right when they need
to wrap.

* Make the organisation switcher more compact

* Remove unused styles

* Update to latest govuk-frontend

* Correction to bundler version

Keep it the same as before.

* Update govuk_tech_docs gem to rebrand the docs