Limit provider checks to production (#5489)
These checks are scheduled to run on Sidekiq nightly in all
environments. They are generate quite a lot of noise in the

The checks are only really useful on `production` so we think it makes
sense to add a conditional to the cron table to only run the job on
production.