[8795] Replace HESA guidance with new API/CSV page (#5528)
* Replace HESA guidance with new API/CSV page

This replaces the _Register trainees through HESA_ page with a new _How
HEIs can Register trainees using the CSV or API functionality_ page.

* [8797] Remove HESA instructions from withdrawal guidance (#5530)

* Remove HESA instructions from withdrawal guidance

This paragraph is no longer relevant

* [8796] Update content on census sign-off guidance page (#5531)

* Update content on census sign-off guidance page

The sign-off form is now built into the service and the _Before you
start_ section is now redundant.

* Update _Bulk upload placement data_ guidance page (#5533)

The HESA references are no longer relevant and the dates need to be
generalised.

* Update app/views/guidance/registering_trainees_through_api_or_csv.md

Co-authored-by: Serafeim Maroulis <serafeim@hey.com>

---------

Co-authored-by: Serafeim Maroulis <serafeim@hey.com>