Simple solution for case-insensitive headers?
This just normalises the headers (converts them to standard casing)
immediately after upload, just before we validate them. The processing
pipeline that follows the upload is unchanged, it can continue to assume
that the headers will be normal case.

e.g. `PROVIDER TRAINEE ID` is converted to `Provider Trainee ID`
and `training route` is converted to `Training Route`.