[8837] Simple solution for case-insensitive headers? (#5570)
* Simple solution for case-insensitive headers?

This just normalises the headers (converts them to standard casing)
immediately after upload, just before we validate them. The processing
pipeline that follows the upload is unchanged, it can continue to assume
that the headers will be normal case.

e.g. `PROVIDER TRAINEE ID` is converted to `Provider Trainee ID`
and `training route` is converted to `Training Route`.

* Appease Rubocop

* Update app/services/bulk_update/add_trainees/import_rows.rb

Co-authored-by: Serafeim Maroulis <serafeim@hey.com>

* Update app/forms/bulk_update/bulk_add_trainees_upload_form.rb

Co-authored-by: Serafeim Maroulis <serafeim@hey.com>

* Appease Rubocop and fix spec regression

* Fix spec regression

---------

Co-authored-by: Serafeim Maroulis <serafeim@hey.com>