Use recommended date for `holdsFrom` when outcome date nil
When we use the TRS API to update trainees professional status we are
not sending a `holdsFrom` date when the state changes to
`recommended_for_award` because we are using the `outcome_date` to populated `holdsFrom`. This is a bit inconsistent with the state
mappings since Register's `recommended_for_award` state maps to the
`Holds` state in TRS. We can pass the `recommended_for_award_at` date
when `outcome_date` is nil.