[8909] Standardise code block markup in tech docs (#5665)
* Use redcarpet and fenced code blocks

This renders triple backticks as expected (as a
`<pre><code>{contents}</code></pre>`). Triple backticks are not part of
standard markdown, rather they are 'Github flavoured markdown'. This is
supported by redcarpet but no the default kramdown markdown parser.

* Convert HTTP request snippets to `pre` blocks

These were using inline `code` blocks which don't render as `pre` blocks
when processed by middleman/redcarpet.

* Re-enable HTML tables in tech docs markdown

This is an option in redcarpet and it's off by default. We have tables
so we need to enable the option.

* Remove trailing blank lines in example code blocks

* Revert back to `kramdown` as the markdown engine

Use indentation for code blocks (as this is standard markdown).

* Remove redundant config

This isn't needed any longer as it's only relevant to the `redcarpet`
markdown engine.

* Missing conversion