Fix flakey JS test for editing schools
This happens very rarely, only when the randomly selected school name is close to
another pre-existing school, e.g. Mallowtown versus Mallowpond.

What seems to be happening is that we are selecting a school that
matches the first few characters of the required name before the
autocomplete has finished filtering. This can lead to the 'wrong' school
being selected and the subsequent assertion failing because the school
names are not an exact match.